+++ ovos_plugin_common_play/ocp/mycroft_cps.py
@@ -3,6 +3,8 @@ from datetime import timedelta
 from os.path import abspath
 
 from mycroft_bus_client.message import dig_for_message
+
+from ovos_plugin_common_play.ocp.constants import OCP_ID
 from ovos_utils.messagebus import Message, wait_for_reply
 
 from ovos_plugin_common_play.ocp.base import OCPAbstractComponent
@@ -55,7 +57,7 @@ class MycroftAudioService:
         # at this stage source == skills, lets indicate audio service took over
         sauce = msg.context.get("source")
         if sauce == "skills":
-            msg.context["source"] = "ovos.common_play"
+            msg.context["source"] = OCP_ID
         return msg
 
     def queue(self, tracks=None):
