+++ ovos_plugin_common_play/ocp/player.py
@@ -14,7 +14,7 @@ from ovos_plugin_common_play.ocp.search import OCPSearch
 from ovos_plugin_common_play.ocp.status import *
 from ovos_plugin_common_play.ocp.mycroft_cps import MycroftAudioService
 from ovos_workshop import OVOSAbstractApplication
-from ovos_plugin_common_play.ocp.base import OCP_ID
+from ovos_plugin_common_play.ocp.constants import OCP_ID
 
 
 class OCPMediaPlayer(OVOSAbstractApplication):
